init: dynamic class/router import managed with importlib.import_module
