+++ app/api/routers.py
@@ -1,8 +1,15 @@
 from fastapi import APIRouter
 
-from app.api.endpoints import SUM, health#, NER, KEY, QA
+import importlib
+from app.core.config import settings
+from app.api.endpoints import health
 
-api_router = APIRouter()
+def create_router(model_type: str) -> APIRouter:
+    model = importlib.import_module(f"app.api.endpoints.{model_type}")
 
-api_router.include_router(SUM.router, prefix="/sum", tags=["summary"])
-api_router.include_router(health.router, prefix="/alive", tags=["health"])
+    api_router = APIRouter()
+    
+    api_router.include_router(model.router, prefix="/run", tags=["execution"])
+    api_router.include_router(health.router, prefix="/alive", tags=["health"])
+    
+    return api_router
